Master volume slider
Since you can import whole playlists from JSON, it gets tedious to
change the volume of 50+ tracks if all of them need to be turned up or
down. This change introduces a mater volume slider on top of the
playlists, that can be enabled with a checkbox next to it. When you
change the master volume slider, the volume for all the tracks changes
accordingly.