Remote libre audio track tab
This adds a new tab in the Tracks modal that takes the playlist JSONs
from DMsGuild201, takes all the tracks that are marked as My Audio and
creates a list of importable tracks. Right now it's useless if you want
to import a whole playlist, but for finding tracks that have already
been uploaded using libre audio, this should help.